[refactor] use new pkg layout, add socketio requirements
