+++ seabus/common/errors.py
