+++ seabus/requirements.txt
@@ -1,18 +1,31 @@
+alembic==0.8.8
 click==6.6
 colorama==0.2.5
 decorator==4.0.10
 Flask==0.11.1
+Flask-Migrate==2.0.0
+Flask-Script==2.0.5
+Flask-SocketIO==2.7.1
+Flask-SQLAlchemy==2.1
 folium==0.2.1
 html5lib==0.999
 itsdangerous==0.24
 Jinja2==2.8
 libais==0.16
+Mako==1.0.4
 MarkupSafe==0.23
 meld3==0.6.10
+numpy==1.11.2
 oboe==1.5.10
+pandas==0.19.0
 pymemcache==1.3.8
+python-dateutil==2.5.3
+python-editor==1.0.1
+python-engineio==1.0.3
+python-socketio==1.6.0
+pytz==2016.7
 requests==2.2.1
-six==1.5.2
+six==1.10.0
 SQLAlchemy==1.0.12
 supervisor==3.0b2
 urllib3==1.7.1
